Example of using JSF.
JSF´s Controller is a class with properties (properties which we link from the VIEW). When we modifies those properties in the Controller, will be modifies in the VIEW as well. So we creates methods (get, set and so on) for modifying them.
Pd: The constructor´s class defines value of properties at the beginning (load the page).